requirements.txt
Flask==2.3.3
pyjwt==2.8.0
requests==2.31.0
protobuf==4.25.1
pycryptodome==3.19.0